ENGR00172185 Auto focus frame is still show even after finish taking a picture.
The root cause was the camera preview region was a hole when using hardware
composer. In the latest code, the hole was cleared after it drawed. The
focus UI was displayed in the hole region. When it should disappear, the focus
UI should be covered by other things. But the hole was cleared after drawed,
there were no other things to cover the focus UI.
The way to fix it, cleared the preview region when there was UI displayed.

Signed-off-by: YangYonggang <b31664@freescale.com>