ENGR00215174 fix hwrotation impact HDMI display problem.
framework is designed to support hwrotation with 270 degree
to make system to be portrait mode by default.

config_reverseDefaultRotation variable in
frameworks/base/core/res/res/values/config.xml
can change the default behavior.
the default value is false.

if the hwrotation is adjusted to 90 degree, the
config_reverseDefaultRotation should be also changed to true
to make system to be portrait mode.
which looks like:
<bool name="config_reverseDefaultRotation">true</bool>

Signed-off-by: Xiaowen Liu <b37945@freescale.com>